Created a full routing system
The pages are now separated and can be accessed separately. The file upload page is alone and the socialbot page is alone. They do not depend on one another and are routed using the BrowserRouter
