+++ client/src/App.css
@@ -13,7 +13,7 @@
 
 .chat-box {
   flex: 1;
-  overflow-y: auto;
+  /* overflow-y: auto; */
   margin-bottom: 20px;
   padding: 10px;
   border: 1px solid #ddd;
