--- client/src/fileUpload.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-function FileUpload() {
-  return (
-    <div>
-      <h1>File Upload Page</h1>
-      <p>This is the file upload page</p>
-    </div>
-  );
-}
-
-export default FileUpload;
