+++ client/src/pages/pages.css
@@ -1,0 +1,61 @@
+.chat-container {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    height: 100vh;
+    max-width: 600px;
+    margin: 0 auto;
+    padding: 20px;
+    background-color: #f5f5f5;
+    border-radius: 10px;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  }
+  
+  .chat-box {
+    flex: 1;
+    /* overflow-y: auto; */
+    margin-bottom: 20px;
+    padding: 10px;
+    border: 1px solid #ddd;
+    background-color: white;
+    border-radius: 10px;
+  }
+  
+  .chat-bubble {
+    margin: 10px 0;
+    padding: 10px 15px;
+    border-radius: 15px;
+    max-width: 70%;
+    word-wrap: break-word;
+  }
+  
+  .user-bubble {
+    align-self: flex-end;
+    justify-self: right;
+    background-color: #d1f4ff;
+    color: #005f8b;
+  }
+  
+  
+  .ai-bubble {
+    align-self: flex-start;
+    background-color: #e4e6eb;
+    color: #333;
+  }
+  
+  
+  textarea {
+    width: 100%;
+    margin-bottom: 10px;
+  }
+  
+  button {
+    display: inline-block;
+    padding: 10px 15px;
+    background-color: #6200ea;
+    color: white;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    transition: background-color 0.3s;
+  }
